Откорректировал упражнение 19 для абстрактной фабрики
