+++ src/main/java/ru/kulichenkom/ekkel/chapter9/game/GameWithAbstractFactory/AbstractGameFactory.java
@@ -3,10 +3,9 @@
 //Конкретная фабрика сразу генерирует продукты. В моем случае, если создать в абстрактной фабрике
 // метод play(), и от него имплементировать фабрики: "подвижные игры", "настольные игры". Эти
 //фабрики уже будут генерироваться в зависимости от дальнейшего кода.
-package ru.kulichenkom.ekkel.chapter9.game.interfaces;
-
-import ru.kulichenkom.ekkel.chapter9.game.TypeOfGame;
+package ru.kulichenkom.ekkel.chapter9.game.GameWithAbstractFactory;
 
 public interface AbstractGameFactory {
-    Game makeGame(TypeOfGame typeOfGame);
+    Game getGame();
 }
+
