+++ src/main/java/ru/kulichenkom/ekkel/chapter9/game/GameWithAbstractFactory/CoinGameFactory.java
@@ -1,0 +1,8 @@
+package ru.kulichenkom.ekkel.chapter9.game.GameWithAbstractFactory;
+
+public class CoinGameFactory implements AbstractGameFactory {
+    @Override
+    public Game getGame() {
+        return new CoinGame();
+    }
+}
