+++ src/main/java/ru/kulichenkom/ekkel/chapter9/game/GameWithAbstractFactory/DiceGameFactory.java
@@ -1,0 +1,8 @@
+package ru.kulichenkom.ekkel.chapter9.game.GameWithAbstractFactory;
+
+public class DiceGameFactory implements AbstractGameFactory {
+    @Override
+    public Game getGame() {
+        return new DiceGame();
+    }
+}
