+++ src/main/java/ru/kulichenkom/ekkel/chapter9/game/GameWithAbstractFactory/Game.java
@@ -1,6 +1,6 @@
 //Ch9Ex19 Создайте на базе паттерна "Фабричный метод" программную среду,
 //моделирующую броски монет и броски кубиков.
-package ru.kulichenkom.ekkel.chapter9.game.interfaces;
+package ru.kulichenkom.ekkel.chapter9.game.GameWithAbstractFactory;
 
 public interface Game {
     void tossUp();
